requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
Pillow               # for ImageField support
gunicorn>=20.0
whitenoise>=6.0
